qader_backend/requirements/development.txt
-r base.txt

django-debug-toolbar==5.1.0
colorama==0.4.6
coverage==7.8.0
factory_boy==3.3.3
Faker==37.1.0
iniconfig==2.1.0
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.11.1